John santos

+++ src/main/java/com/codeup/springblog/controllers/PostController.java
@@ -1,23 +1,28 @@
 package com.codeup.springblog.controllers;
 
 import com.codeup.springblog.models.Post;
+import com.codeup.springblog.models.User;
 import com.codeup.springblog.repositories.PostRepo;
-import org.hibernate.service.spi.InjectService;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.codeup.springblog.repositories.UserRepo;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
-import org.thymeleaf.model.IModel;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
-import java.util.ArrayList;
+import java.security.Principal;
 
 @Controller
 public class PostController {
 
     private PostRepo postDao;
+    private UserRepo userDao;
 
-    public PostController(PostRepo postDao) {
+    public PostController(PostRepo postDao, UserRepo userDao) {
         this.postDao = postDao;
+        this.userDao = userDao;
     }
 
     @GetMapping("/posts")
@@ -27,29 +32,47 @@ public class PostController {
     }
 
     @GetMapping("/posts/{id}")
-    public String getPost(@PathVariable int id, Model model){
-        Post post1 = new Post(id, "Europa's First Post", "Remote Learning Today!");
-        model.addAttribute("title", post1.getTitle());
-        model.addAttribute("body", post1.getBody());
+    public String getPost(@PathVariable long id, Model model, Principal principal){
+//        User loggedIn = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String userName = "";
+        if (principal != null) {
+            userName = principal.getName();
+//            userDao.findByUsername(userName);
+        }
+        model.addAttribute("userName", userName);
+        model.addAttribute("post",postDao.getOne(id));
         return "posts/show";
     }
 
     @GetMapping("/posts/create")
-    @ResponseBody
     public String getCreatePostForm(){
-        return "view the form for creating a post";
+        User loggedIn = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (loggedIn != null)
+            return "posts/create";
+        else
+            return "redirect:/login";
     }
 
     @PostMapping("/posts/create")
-    @ResponseBody
-    public String createPost(){
-        return "create a new post";
+    public String createPost(@RequestParam String title, @RequestParam String body ){
+        Post newPost = new Post();
+        newPost.setTitle(title);
+        newPost.setBody(body);
+        User loggedIn = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        newPost.setUser(loggedIn);
+        postDao.save(newPost);
+        return "redirect:/posts";
     }
 
     @PostMapping("/posts/{id}/delete")
     public String delete(@PathVariable long id){
+        System.out.println((User) SecurityContextHolder.getContext().getAuthentication().getPrincipal());
+        User loggedInUser = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (loggedInUser.getId() == postDao.getOne(id).getUser().getId())
         // delete post
-        postDao.deleteById(id);
+            postDao.deleteById(id);
+
         return "redirect:/posts";
     }
